+++ src/api/API.tsx
@@ -21,19 +21,26 @@ export const reissueToken = async (
   return data;
 };
 
+const errorMapping: Record<number, { emoji: string; translationKey: string }> = {
+  2011: { emoji: '🚫', translationKey: 'feed:error.alreadyBlocked' },
+  4000: { emoji: '🗑️', translationKey: 'feed:error.deletedFeed' },
+  4006: { emoji: '🗑️', translationKey: 'feed:error.deletedComment' },
+  5004: { emoji: '📩', translationKey: 'feed:error.alreadyExistChatRequest' },
+  5005: { emoji: '💬', translationKey: 'feed:error.alreadyExistChatroom' },
+};
+
 API.interceptors.response.use(
   (response) => response,
   async (error) => {
     if (!error.response) {
       useToastStore.getState().showToast(<MyText>🌐</MyText>, i18n.t('common:toast.error.network'));
-      logError(error);
       return Promise.reject(error);
     }
     const errorCode = error.response?.data?.code;
+
     if (errorCode === 9000) {
       useToastStore.getState().showToast(<MyText>⚠️</MyText>, error.response.data.message, 2000);
-    }
-    if (error.response?.status === 401 && errorCode === 3002 && !error.config._retry) {
+    } else if (error.response?.status === 401 && errorCode === 3002 && !error.config._retry) {
       error.config._retry = true;
       try {
         const refreshToken = await SecureStore.getItemAsync(TOKEN_KEYS.REFRESH);
@@ -52,10 +59,20 @@ API.interceptors.response.use(
         showErrorModal('tokenExpired');
         return Promise.reject(reissueError);
       }
+    } else {
+      const errorInfo = errorMapping[errorCode];
+      if (errorInfo) {
+        useToastStore
+          .getState()
+          .showToast(<MyText>{errorInfo.emoji}</MyText>, i18n.t(errorInfo.translationKey), 2000);
+      } else {
+        useToastStore
+          .getState()
+          .showToast(<MyText>⚠️</MyText>, i18n.t('common:toast.error.unknown'), 2000);
+      }
     }
-    useToastStore
-      .getState()
-      .showToast(<MyText>⚠️</MyText>, i18n.t('common:toast.error.unknown'), 2000);
+
+    logError(error);
     return Promise.reject(error);
   }
 );
