+++ src/components/modal/BlockModal.tsx
@@ -2,6 +2,7 @@ import { useTranslation } from 'react-i18next';
 import { Text } from 'react-native';
 import { UserRepository } from '@/api';
 import { useToastStore } from '@/store';
+import { logError } from '@/utils';
 import { MyText } from '../common';
 import { StandardModal } from './Common';
 
@@ -24,16 +25,7 @@ export function BlockModal({ visible, buddyId, onClose, onBlockSuccess }: BlockM
       }
       showToast(<MyText>🚫</MyText>, t('toast.blockSuccess'), 2000);
     } catch (error: any) {
-      const errorCode = error.response?.data?.code;
-      const errorMapping: Record<number, { emoji: string; translationKey: string }> = {
-        2011: { emoji: '🚫', translationKey: 'feed:error.alreadyBlocked' },
-        4000: { emoji: '🗑️', translationKey: 'feed:error.deletedFeed' },
-        4006: { emoji: '🗑️', translationKey: 'feed:error.deletedComment' },
-      };
-      const errorInfo = errorMapping[errorCode];
-      if (errorInfo) {
-        showToast(<Text>{errorInfo.emoji}</Text>, t(errorInfo.translationKey), 2000);
-      }
+      logError(error);
     }
   };
