+++ src/components/modal/ReportModal.tsx
@@ -3,6 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { Alert, TextInput, Text } from 'react-native';
 import { UserRepository } from '@/api';
 import { useToastStore } from '@/store';
+import { logError } from '@/utils';
 import { StandardModal } from './Common';
 
 interface ReportModalProps {
@@ -43,15 +44,7 @@ export function ReportModal({
       }
       showToast(<Text>🙌</Text>, t('toast.reportSuccess'), 2000);
     } catch (error: any) {
-      const errorCode = error.response?.data?.code;
-      const errorMapping: Record<number, { emoji: string; translationKey: string }> = {
-        4000: { emoji: '🗑️', translationKey: 'feed:error.deletedFeed' },
-        4006: { emoji: '🗑️', translationKey: 'feed:error.deletedComment' },
-      };
-      const errorInfo = errorMapping[errorCode];
-      if (errorInfo) {
-        showToast(<Text>{errorInfo.emoji}</Text>, t(errorInfo.translationKey), 2000);
-      }
+      logError(error);
     } finally {
       handleCancel();
     }
